+++ classes/Mail/TransportFactoryException.php
@@ -1,0 +1,4 @@
+<?php
+namespace AppZap\PHPFramework\Mail;
+
+class TransportFactoryException extends \Exception {}